Move class AvatarColorsProvider to its own file.
