+++ libraries/designsystem/src/main/kotlin/io/element/android/libraries/designsystem/colors/AvatarColors.kt
@@ -16,50 +16,10 @@
 
 package io.element.android.libraries.designsystem.colors
 
-import androidx.collection.LruCache
 import androidx.compose.ui.graphics.Color
-import io.element.android.libraries.theme.colors.avatarColorsDark
-import io.element.android.libraries.theme.colors.avatarColorsLight
 
 data class AvatarColors(
     val background: Color,
     val foreground: Color,
 )
-
-object AvatarColorsProvider {
-    private val cache = LruCache<String, AvatarColors>(200)
-    private var currentThemeIsLight = true
-
-    fun provide(id: String, isLightTheme: Boolean): AvatarColors {
-        if (currentThemeIsLight != isLightTheme) {
-            currentThemeIsLight = isLightTheme
-            cache.evictAll()
-        }
-        val valueFromCache = cache.get(id)
-        return if (valueFromCache != null) {
-            valueFromCache
-        } else {
-            val colors = avatarColors(id, isLightTheme)
-            cache.put(id, colors)
-            colors
-        }
-    }
-
-    private fun avatarColors(id: String, isLightTheme: Boolean): AvatarColors {
-        val hash = id.toHash()
-        val colors = if (isLightTheme) {
-            avatarColorsLight[hash]
-        } else {
-            avatarColorsDark[hash]
-        }
-        return AvatarColors(
-            background = colors.first,
-            foreground = colors.second,
-        )
-    }
-}
-
-internal fun String.toHash(): Int {
-    return toList().sumOf { it.code } % avatarColorsLight.size
-}
